Move AbsoluteSyntaxInfo to "edge" SyntaxData information
The root node doesn't need to hold its `AbsoluteSyntaxInfo` because
it's always zero. Move this information to "edge" data so only non-root
nodes hold the "absolute info". Introduce an "edge" info struct, and use
it with `indirect` enum case instead of `SyntaxBox`.

Previously the structure of `SyntaxData` was:
 * AbsoluteRawSyntax
   * RawSyntax
   * AbsoluteSyntaxInfo
 * Tagged union of
   * SyntaxArena
   * SyntaxBox
     * Parent SyntaxData

Now:
 * RawSyntax node
 * Tagged union of:
   * SyntaxData.Info.Root
     * SyntaxArena
   * SyntaxData.Info.Edge (indirect case)
     * Parent SyntaxData
     * AbsoluteSyntaxInfo

This reduces the size of `SyntaxData` to 2 words.

Also `SyntaxData.parent` now returns `SyntaxData` instead of `Syntax`.
`SyntaxData` should not rely on `Syntax` API at all.