Eliminate `SyntaxNode`
`SyntaxNode` was introduced as a lighter `Syntax` when existentials were
needed for creating `Syntax` instance.

Now that `Syntax`/`SyntaxData` is just a struct that holds one class
(`SyntaxBox`). So it's not that heavy.
More importantly `SyntaxNode` was not safe to escape because it didn't
retain the `SyntaxArena` of the `AbsoluteRawSyntax`

`SyntaxNode` was only used in `SyntaxCursor` for `IncrementalParseLookup`.
Use `SyntaxData` instead.