Simplify SyntaxClassifications implementation
Previously, `SyntaxClassifications` was a lazy sequence backed by an
iterator that traverses the tree. While it works perfectly, walking in a
tree strucuture using an interator was not a simple task - `next()`
operation may walk into child nodes, find a sibling, or find an
ancestors sibling, while skipping out-of-range nodes, etc. Also this
classfication iterator was nested, token iterator and an iterator inside
a token. In addition to that it has a buffering mechanism to merge
consecutive classification ranges of the same kind.

This change simplify the implementaiton by using a recusive visitor. It
uses some more memory as it now eagarly constructs an array, but the
performance is improved